refactor: cleanup entry point
